split Backbone.js into a bunch of different files. Still more work to do to modularize it
