+++ frontend/src/Metamaps/DataModel/MappingCollection.js
@@ -1,0 +1,11 @@
+import Backbone from 'backbone'
+Backbone.$ = window.$
+
+import Mapping from './Mapping'
+
+const MappingCollection = Backbone.Collection.extend({
+  model: Mapping,
+  url: '/mappings'
+})
+
+export default MappingCollection
