+++ frontend/src/Metamaps/DataModel/MetacodeCollection.js
@@ -1,0 +1,16 @@
+import Backbone from 'backbone'
+Backbone.$ = window.$
+
+import Metacode from './Metacode'
+
+const MetacodeCollection = Backbone.Collection.extend({
+  model: Metacode,
+  url: '/metacodes',
+  comparator: function (a, b) {
+    a = a.get('name').toLowerCase()
+    b = b.get('name').toLowerCase()
+    return a > b ? 1 : a < b ? -1 : 0
+  }
+})
+
+export default MetacodeCollection
